app/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
scikit-learn==0.24.2
nltk==3.6.2
pandas==1.3.0
requests==2.26.0
joblib==1.0.1

fastapi
requests
python-dotenv
uvicorn
